Move offline entities code to module
